+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -100,7 +100,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-16T07:22:48.465Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-16T09:24:08.283Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")