+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXMWXR85D5ZE4BRXA071JJYZ"
+        "01KXN3VS7KWX7WXBZ9J1R8ZYZG"
     )
     r"""The unique identifier of the data source"""