+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsehttp.py
@@ -55,9 +55,9 @@ class EvaluatorResponseHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-16T07:22:50.133Z"
+    created: Optional[str] = "2026-07-16T09:24:13.490Z"
 
-    updated: Optional[str] = "2026-07-16T07:22:50.133Z"
+    updated: Optional[str] = "2026-07-16T09:24:13.490Z"
 
     updated_by_id: OptionalNullable[str] = UNSET