+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXMWXR81YC09DWWP05KEXXNK"
+        "01KXN3VS7F0017FZ77W4WMWVW4"
     )
     r"""The unique identifier of the data source"""