+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXMWXR84R9XAV6VRJ9WHMQ0X"
+        "01KXN3VS7HJ0F0SAM0EB7Q6PSC"
     )
     r"""The unique identifier of the data source"""