+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2012,7 +2012,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KXMWXQSJXHKDHA1W5WCCZ2MP"
+    id: Optional[str] = "01KXN3VRQFZ73FCKQGSHX03CF6"
 
     description: Optional[str] = None